// src/app/core/api/models/mapa-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MapaData } from '../models/mapa-data';

/**
 * Resposta do mapa: descrição + lista de pontos georreferenciados.
 */
export interface MapaResponse {
  data: MapaData;

  /**
   * Identificador semântico do mapa.
   */
  descricao: string;
}