// src/app/core/api/models/painel-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PainelData } from '../models/painel-data';

/**
 * Resposta do painel: descrição + bloco de dados com gráficos e filtros.
 */
export interface PainelResponse {
  data: PainelData;

  /**
   * Identificador semântico do painel.
   */
  descricao: string;
}